+++ include/YODA/Profile1D.h
@@ -273,8 +273,8 @@ namespace YODA {
 
     /// @todo Add integrals? Or are they too ambiguous to make a core function?
 
-    /// Get the number of fills
-    unsigned long numEntries(bool includeoverflows=true) const;
+    /// Get the number of fills (fractional fills are possible)
+    double numEntries(bool includeoverflows=true) const;
 
     /// Get the effective number of fills
     double effNumEntries(bool includeoverflows=true) const;
