+++ src/Histo2D.cc
@@ -50,7 +50,7 @@ namespace YODA {
 
   /////////////// COMMON TO ALL BINNED
 
-  unsigned long Histo2D::numEntries(bool includeoverflows) const {
+  double Histo2D::numEntries(bool includeoverflows) const {
     if (includeoverflows) return totalDbn().numEntries();
     unsigned long n = 0;
     for (const Bin& b : bins()) n += b.numEntries();
