+++ src/Profile1D.cc
@@ -46,7 +46,7 @@ namespace YODA {
 
   /////////////// COMMON TO ALL BINNED
 
-  unsigned long Profile1D::numEntries(bool includeoverflows) const {
+  double Profile1D::numEntries(bool includeoverflows) const {
     if (includeoverflows) return totalDbn().numEntries();
     unsigned long n = 0;
     for (const Bin& b : bins()) n += b.numEntries();
